trying script in different spot
